[Event] Move Data class to Data namespace.
