+++ src/Valkyrja/Application/Data/Data.php
@@ -15,7 +15,7 @@ namespace Valkyrja\Application\Data;
 
 use Valkyrja\Cli\Routing\Data\Data as CliData;
 use Valkyrja\Container\Data\Data as ContainerData;
-use Valkyrja\Event\Data as EventData;
+use Valkyrja\Event\Data\Data as EventData;
 use Valkyrja\Http\Routing\Data as HttpData;
 
 /**
