+++ src/Valkyrja/Application/Valkyrja.php
@@ -22,7 +22,7 @@ use Valkyrja\Application\Support\Component;
 use Valkyrja\Cli\Routing\Data\Data as CliData;
 use Valkyrja\Container\Data\Data as ContainerData;
 use Valkyrja\Container\Manager\Contract\Container;
-use Valkyrja\Event\Data as EventData;
+use Valkyrja\Event\Data\Data as EventData;
 use Valkyrja\Http\Routing\Data as HttpData;
 
 /**
