+++ src/Valkyrja/Event/Collection/Collection.php
@@ -15,8 +15,8 @@ namespace Valkyrja\Event\Collection;
 
 use Override;
 use Valkyrja\Event\Collection\Contract\Collection as Contract;
-use Valkyrja\Event\Data;
 use Valkyrja\Event\Data\Contract\Listener;
+use Valkyrja\Event\Data\Data;
 use Valkyrja\Event\Data\Listener as Model;
 use Valkyrja\Event\Exception\InvalidArgumentException;
