+++ src/Valkyrja/Event/Data/Data.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Event;
+namespace Valkyrja\Event\Data;
 
 use Valkyrja\Event\Data\Contract\Listener;
